# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Todo App API",
    "version": "1.0.0",
    "description": "API documentation for the Todo App."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Creates a new user account and sends a welcome email.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully!"
          },
          "400": {
            "description": "Validation error or registration failed."
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "description": "Authenticates the user and returns a JWT token.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful, returns JWT token."
          },
          "400": {
            "description": "Invalid credentials."
          }
        }
      }
    },
    "/todos": {
      "post": {
        "summary": "Create a new todo",
        "description": "Creates a new todo item for the authenticated user.",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "Bearer token for authentication."
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Buy groceries"
                },
                "description": {
                  "type": "string",
                  "example": "Milk, eggs, bread"
                },
                "priority": {
                  "type": "string",
                  "enum": ["low", "medium", "high"],
                  "example": "medium"
                },
                "dueDate": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2024-12-01T12:00:00Z"
                }
              },
              "required": ["title"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Todo created successfully!"
          },
          "400": {
            "description": "Validation error or creation failed."
          }
        }
      },
      "get": {
        "summary": "Get todos",
        "description": "Retrieves todos for the authenticated user with optional pagination, filtering, and search.",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "Bearer token for authentication."
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "default": 1
          },
          {
            "in": "query",
            "name": "limit",
            "type": "integer",
            "default": 10
          },
          {
            "in": "query",
            "name": "status",
            "type": "string",
            "description": "Filter by status."
          },
          {
            "in": "query",
            "name": "priority",
            "type": "string",
            "description": "Filter by priority."
          },
          {
            "in": "query",
            "name": "search",
            "type": "string",
            "description": "Search todos by title or description."
          }
        ],
        "responses": {
          "200": {
            "description": "Todos retrieved successfully."
          }
        }
      }
    }
  },
  "definitions": {},
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token for authorization."
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}